Json pimp

+++ Hero.cpp
@@ -2,8 +2,8 @@
 #include "JSON.h"
 #include <iostream>
 
-Hero::Hero(const std::string &name, const int health, const int damage, const float atkcooldown,
-           const int ExperiencePerLevel, const int HealthPointBonusPerLevel, const int DamageBonusPerLevel, const float ColdownMultiplierPerLevel)
+Hero::Hero(const std::string &name, const int health, const int damage, const double atkcooldown,
+           const int ExperiencePerLevel, const int HealthPointBonusPerLevel, const int DamageBonusPerLevel, const double ColdownMultiplierPerLevel)
     : Monster{name, health, damage, atkcooldown}, maxHealth(health), ExperiencePerLevel(ExperiencePerLevel), HealthPointBonusPerLevel(HealthPointBonusPerLevel),
       DamageBonusPerLevel(DamageBonusPerLevel), ColdownMultiplierPerLevel(ColdownMultiplierPerLevel)
 {
@@ -56,9 +56,9 @@ Hero Hero::parse(const std::string &fileName)
         data.get<std::string>("name"),
         data.get<int>("base_health_points"),
         data.get<int>("base_damage"),
-        data.get<float>("base_attack_cooldown"),
+        data.get<double>("base_attack_cooldown"),
         data.get<int>("experience_per_level"),
         data.get<int>("health_point_bonus_per_level"),
         data.get<int>("damage_bonus_per_level"),
-        data.get<float>("cooldown_multiplier_per_level"));
+        data.get<double>("cooldown_multiplier_per_level"));
 }
